Assert the unavailable branch on both flavors instead of skipping it
The three *UnsupportedTest classes reached the detailed-tracking-
unavailable state through Assume, which JUnit reports as a skip. On
classic — which declares specialUse and so is never blocked — six of
their eight methods therefore asserted nothing at all. A branch covered
on one flavor is worse than the flavor split CLAUDE.md describes: there
the two flavors assert different things, here one asserted nothing.

The sibling classes were already doing it the right way, at the same
sdk 34 with the same denyPermissions, by comparing against the constant:
assertEquals(Distribution.HAS_SPECIAL_USE_FGS, ...). Going through the
three classes assertion by assertion, everything in them was either
already covered there or folds into it, so all three are deleted.

Three assertions were genuinely new and moved rather than vanished. The
detailed-tracking switch is now asserted dead beside the note that
explains it, since a live switch would spring back on every tap and read
as a bug rather than a limit. The refusal test asserts no service was
started behind a setting that was never written. And the case where the
setting arrives from Android's own backup on a build that cannot act on
it — never granted rather than revoked mid-flight, which is what an
actual device restore looks like — is a new test in TrackingControllerTest.
That one now advances both clocks rather than the wall clock alone:
moving wall time by itself drags bootAt with it and would have closed the
session for pre-dating the boot, so it could have passed for the wrong
reason.

Also three comments whose reason was deleted with the code around them:
a dangling KDoc line in TrackingService, PlaybackWatcher's remaining
statement of what rather than why, and the IcuUnits KDoc, which narrated
the null-gate that is gone instead of the reasons that still stand.
